Labs: Navigating to Instructor java examples #206   - Renamed the classes for this lesson
