+++ cross-site-scripting/src/main/java/org/owasp/webgoat/plugin/crosssitescripting/EditProfileCrossSiteScripting.java
@@ -7,12 +7,11 @@ import org.owasp.webgoat.plugin.GoatHillsFinancial.GoatHillsFinancial;
 import org.owasp.webgoat.session.ParameterNotFoundException;
 import org.owasp.webgoat.session.UnauthenticatedException;
 import org.owasp.webgoat.session.UnauthorizedException;
-import org.owasp.webgoat.session.ValidationException;
 import org.owasp.webgoat.session.WebSession;
 
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
 
 
 /***************************************************************************************************
@@ -42,37 +41,26 @@ import java.sql.Statement;
  * 
  * For details, please see http://webgoat.github.io
  */
-public class ViewProfile extends DefaultLessonAction
+public class EditProfileCrossSiteScripting extends DefaultLessonAction
 {
 
-    public ViewProfile(GoatHillsFinancial lesson, String lessonName, String actionName)
+    public EditProfileCrossSiteScripting(GoatHillsFinancial lesson, String lessonName, String actionName)
     {
         super(lesson, lessonName, actionName);
     }
 
     public void handleRequest(WebSession s) throws ParameterNotFoundException, UnauthenticatedException,
-            UnauthorizedException, ValidationException
+            UnauthorizedException
     {
         getLesson().setCurrentAction(s, getActionName());
 
         if (isAuthenticated(s))
         {
-            int userId = getIntSessionAttribute(s, getLessonName() + "." + CrossSiteScripting.USER_ID);
-            int employeeId = -1;
-            try
-            {
-                // User selected employee
-                employeeId = s.getParser().getIntParameter(CrossSiteScripting.EMPLOYEE_ID);
-            } catch (ParameterNotFoundException e)
-            {
-                // May be an internally selected employee
-                employeeId = getIntRequestAttribute(s, getLessonName() + "." + CrossSiteScripting.EMPLOYEE_ID);
-            }
+            int userId = getUserId(s);
+            int employeeId = s.getParser().getIntParameter(CrossSiteScripting.EMPLOYEE_ID);
 
             Employee employee = getEmployeeProfile(s, userId, employeeId);
             setSessionAttribute(s, getLessonName() + "." + CrossSiteScripting.EMPLOYEE_ATTRIBUTE_KEY, employee);
-
-            updateLessonStatus(s, employee);
         }
         else
             throw new UnauthenticatedException();
@@ -80,7 +68,7 @@ public class ViewProfile extends DefaultLessonAction
 
     public String getNextPage(WebSession s)
     {
-        return CrossSiteScripting.VIEWPROFILE_ACTION;
+        return CrossSiteScripting.EDITPROFILE_ACTION;
     }
 
     public Employee getEmployeeProfile(WebSession s, int userId, int subjectUserId) throws UnauthorizedException
@@ -90,15 +78,16 @@ public class ViewProfile extends DefaultLessonAction
         // Query the database for the profile data of the given employee
         try
         {
-            String query = "SELECT * FROM employee WHERE userid = " + subjectUserId;
+            String query = "SELECT * FROM employee WHERE userid = ?";
+
             try
             {
-                Statement answer_statement = WebSession.getConnection(s)
-                        .createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
-                ResultSet answer_results = answer_statement.executeQuery(query);
+                PreparedStatement answer_statement = WebSession.getConnection(s)
+                        .prepareStatement(query, ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
+                answer_statement.setInt(1, subjectUserId);
+                ResultSet answer_results = answer_statement.executeQuery();
                 if (answer_results.next())
                 {
-
                     // Note: Do NOT get the password field.
                     profile = new Employee(answer_results.getInt("userid"), answer_results.getString("first_name"),
                             answer_results.getString("last_name"), answer_results.getString("ssn"), answer_results
@@ -128,22 +117,19 @@ public class ViewProfile extends DefaultLessonAction
 
     public Employee getEmployeeProfile_BACKUP(WebSession s, int userId, int subjectUserId) throws UnauthorizedException
     {
-        // Query the database to determine if this employee has access to this function
-        // Query the database for the profile data of the given employee if "owned" by the given
-        // user
-
         Employee profile = null;
 
         // Query the database for the profile data of the given employee
         try
         {
-            String query = "SELECT * FROM employee WHERE userid = " + subjectUserId;
+            String query = "SELECT * FROM employee WHERE userid = ?";
 
             try
             {
-                Statement answer_statement = WebSession.getConnection(s)
-                        .createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
-                ResultSet answer_results = answer_statement.executeQuery(query);
+                PreparedStatement answer_statement = WebSession.getConnection(s)
+                        .prepareStatement(query, ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
+                answer_statement.setInt(1, subjectUserId);
+                ResultSet answer_results = answer_statement.executeQuery();
                 if (answer_results.next())
                 {
                     // Note: Do NOT get the password field.
@@ -155,7 +141,6 @@ public class ViewProfile extends DefaultLessonAction
                                     .getInt("salary"), answer_results.getString("ccn"), answer_results
                                     .getInt("ccn_limit"), answer_results.getString("disciplined_date"), answer_results
                                     .getString("disciplined_notes"), answer_results.getString("personal_description"));
-
                     /*
                      * System.out.println("Retrieved employee from db: " + profile.getFirstName() +
                      * " " + profile.getLastName() + " (" + profile.getId() + ")");
@@ -172,43 +157,6 @@ public class ViewProfile extends DefaultLessonAction
         }
 
         return profile;
-    }
-
-    private void updateLessonStatus(WebSession s, Employee employee)
-    {
-        String stage = getStage(s);
-        int userId = -1;
-        try
-        {
-            userId = getIntSessionAttribute(s, getLessonName() + "." + CrossSiteScripting.USER_ID);
-        } catch (ParameterNotFoundException pnfe)
-        {
-        }
-        if (CrossSiteScripting.STAGE1.equals(stage))
-        {
-            String address1 = employee.getAddress1().toLowerCase();
-            if (userId != employee.getId() && address1.indexOf("<script>") > -1 && address1.indexOf("alert") > -1
-                    && address1.indexOf("</script>") > -1)
-            {
-                setStageComplete(s, CrossSiteScripting.STAGE1);
-            }
-        }
-        else if (CrossSiteScripting.STAGE3.equals(stage))
-        {
-            String address2 = employee.getAddress1().toLowerCase();
-            if (address2.indexOf("<script>") > -1 && address2.indexOf("alert") > -1
-                    && address2.indexOf("</script>") > -1)
-            {
-                setStageComplete(s, CrossSiteScripting.STAGE3);
-            }
-        }
-        else if (CrossSiteScripting.STAGE4.equals(stage))
-        {
-            if (employee.getAddress1().toLowerCase().indexOf("&lt;") > -1)
-            {
-                setStageComplete(s, CrossSiteScripting.STAGE4);
-            }
-        }
     }
 
 }
